ReactNode for text for Compositions
