+++ packages/constellation/src/components/Compositions/Alerts/Banner/Banner.types.ts
@@ -6,11 +6,11 @@ export type Props<T extends React.ElementType> = PolymorphicComponentPropsWithRe
     /**
      * The second part of the banner message that is displaed in bold.
      */
-    boldContent?: string
+    boldContent?: string | React.ReactNode
     /**
      * The first part of the banner message that is displayed in regular font weight.
      */
-    content: string
+    content: string | React.ReactNode
     /**
      * The icon displayed after content and before bold content.
      */
