+++ packages/constellation/src/components/Compositions/Table/Cells/InformationalCell/InformationalCell.types.ts
@@ -1,4 +1,4 @@
-import { FC } from 'react'
+import { FC, ReactNode } from 'react'
 
 import { TagProps } from '../../../../Primitives'
 import { BaseCellProps } from '../..'
@@ -7,7 +7,7 @@ interface Props extends BaseCellProps, TagProps {
   /**
    * The text content displayed in the Tag
    */
-  text: string
+  text: string | ReactNode
 }
 type Component = FC<Props>
